+++ component/component.go
@@ -61,8 +61,8 @@ type Component interface {
 	Shutdown(ctx context.Context) error
 }
 
-// ComponentSettings is passed to ReceiverFactory.Create* functions.
-type ComponentSettings struct {
+// Settings is passed to ReceiverFactory.Create* functions.
+type Settings struct {
 	// Logger that the factory can use during creation and can pass to the created
 	// component to be used later as well.
 	Logger *zap.Logger